fixup! node-api: enable uncaught exceptions policy by default
